+++ spec/methods/$afterInvoke.spec.js
@@ -7,14 +7,14 @@ describe("$afterInvoke", function () {
   });
 
   it("should error if no function provided", function () {
-    expect(() => Jpex.$afterInvoke()).toThrow();
+    expect(() => Jpex.$afterInvoke = 'asb').toThrow();
   });
   it("should call a function before invoking", function () {
     var order = [];
     var spy1 = jasmine.createSpy().and.callFake(() => order.push(1));
     var spy2 = jasmine.createSpy().and.callFake(() => order.push(2));
     var Class = Jpex.extend(spy1);
-    Class.$afterInvoke(spy2);
+    Class.$afterInvoke = spy2;
 
     Class();
 
@@ -27,23 +27,23 @@ describe("$afterInvoke", function () {
     var spy = jasmine.createSpy();
     var path = require('path');
     var Class = Jpex.extend({dependencies : 'path'});
-    Class.$afterInvoke(spy);
+    Class.$afterInvoke = spy;
     Class();
 
     expect(spy).toHaveBeenCalledWith(path);
   });
   it("should have the instance context", function () {
     var calledWith = {};
-    Jpex.$afterInvoke(function () {
+    Jpex.$afterInvoke = function () {
       calledWith = this;
-    });
+    };
     Jpex();
 
     expect(calledWith).toBe(Jpex.$instances[0]);
   });
   it("should not fire on child classes", function () {
     var called = false;
-    Jpex.$afterInvoke(() => called = true);
+    Jpex.$afterInvoke = () => called = true;
     var Class = Jpex.extend();
 
     Class();